issue #124: move :providers sources and wire build files
Deviations from spec (all necessary to resolve cross-module compile deps):
- SMART_HOME_OBJECT_MAPPER + Jackson serializer classes extracted from
  JacksonConfiguration.kt to core/src/main/.../jackson/SmartHomeObjectMapper.kt
  so both :providers:main and :persistence can reach it without a cycle;
  fully-qualified name unchanged so all existing imports remain identical.
- Jackson deps (jackson-module-kotlin, jackson-datatype-jsr310) added to
  :core:main where the mapper definition now lives.
- createMockEngine extracted from root src/test/.../controllers/Utils.kt to
  providers/src/testFixtures/.../controllers/MockEngineFactory.kt (same
  package org.agrfesta.sh.api.controllers) so :providers:test can import it
  without a circular dependency; root tests access it via testFixtures(:providers).
- :providers:build.gradle.kts adds java-test-fixtures plugin, resolutionStrategy.force
  for Kotlin stdlib, spring-boot (for @ConfigurationProperties), arrow.core, and
  test/testFixtures dependencies not listed in the spec but required to compile
  the moved sources (kotest-assertions-json, spring-context).
