+++ build.gradle.kts
@@ -23,6 +23,7 @@ repositories {
 dependencies {
 	implementation(project(":core"))
 	implementation(project(":persistence"))
+	implementation(project(":providers"))
 	implementation("org.jetbrains.kotlin:kotlin-reflect")
 	implementation("org.springframework.boot:spring-boot-starter-web")
 	implementation("org.springframework.boot:spring-boot-starter-jdbc")
@@ -41,6 +42,7 @@ dependencies {
 
 	testImplementation(testFixtures(project(":core")))
 	testImplementation(testFixtures(project(":persistence")))
+	testImplementation(testFixtures(project(":providers")))
 	testImplementation("org.springframework.boot:spring-boot-starter-test") {
 		exclude(module = "junit")
 		exclude(module = "junit-vintage-engine")
