+++ providers/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     alias(libs.plugins.jvm)
     alias(libs.plugins.kotlin.spring)
+    `java-test-fixtures`
 }
 
 kotlin {
@@ -14,12 +15,43 @@ repositories {
     mavenCentral()
 }
 
+// Same reason as :core and :persistence — MockK pulls kotlin-stdlib:2.1.x without Boot plugin
+configurations.all {
+    resolutionStrategy.force(
+        "org.jetbrains.kotlin:kotlin-stdlib:1.9.24",
+        "org.jetbrains.kotlin:kotlin-reflect:1.9.24"
+    )
+}
+
 dependencies {
+    implementation(platform(libs.spring.boot.bom))
     implementation(project(":core"))
+    implementation("org.springframework.boot:spring-boot")
+    implementation(libs.arrow.core)
     implementation(libs.ktor.core)
     implementation(libs.ktor.okhttp)
     implementation(libs.ktor.content.negotiation)
     implementation(libs.ktor.jackson)
+
+    testFixturesImplementation(platform(libs.spring.boot.bom))
+    testFixturesImplementation(project(":core"))
+    testFixturesImplementation(testFixtures(project(":core")))
+    testFixturesImplementation("org.springframework:spring-context")
+    testFixturesImplementation(libs.ktor.client.mock)
+    testFixturesImplementation(libs.mockk)
+    testFixturesImplementation(libs.kotest.assertions.core)
+    testFixturesImplementation(libs.kotest.assertions.json)
+
+    testImplementation(project(":core"))
+    testImplementation(testFixtures(project(":core")))
+    testImplementation("org.junit.jupiter:junit-jupiter-api")
+    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
+    testRuntimeOnly("org.junit.platform:junit-platform-launcher")
+    testImplementation(libs.kotest.assertions.core)
+    testImplementation(libs.kotest.assertions.arrow)
+    testImplementation(libs.kotest.assertions.json)
+    testImplementation(libs.mockk)
+    testImplementation(libs.ktor.client.mock)
 }
 
 tasks.withType<Test> {
