+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/netatmo/NetatmoModel.kt
