+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/SwitchBotConfiguration.kt
