+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/SwitchBotService.kt
