+++ providers/src/main/kotlin/org/agrfesta/sh/api/providers/switchbot/devices/SwitchBotMiniHub.kt
