+++ providers/src/test/kotlin/org/agrfesta/sh/api/providers/netatmo/devices/NetatmoSmartherTest.kt
