+++ providers/src/test/kotlin/org/agrfesta/sh/api/providers/switchbot/SwitchBotDevicesClientTest.kt
