+++ providers/src/testFixtures/kotlin/org/agrfesta/sh/api/providers/netatmo/BehaviorRegistry.kt
