+++ providers/src/testFixtures/kotlin/org/agrfesta/sh/api/providers/switchbot/SwitchBotResponsesMother.kt
