+++ providers/src/testFixtures/resources/org/agrfesta/sh/api/providers/netatmo/resp.json
