+++ src/test/kotlin/org/agrfesta/sh/api/controllers/Utils.kt
@@ -1,10 +1,7 @@
 package org.agrfesta.sh.api.controllers
 
-import io.ktor.client.engine.mock.MockEngine
-import io.ktor.client.engine.mock.respond
 import io.restassured.specification.RequestSpecification
 import org.agrfesta.sh.api.core.domain.devices.Device
-import org.agrfesta.sh.api.providers.netatmo.BehaviorRegistry
 import org.agrfesta.test.mothers.aRandomUniqueString
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder
 
@@ -25,13 +22,3 @@ fun RequestSpecification.authenticated(): RequestSpecification =
 
 fun RequestSpecification.wrongAuthentication(): RequestSpecification =
     header("Authorization", "Bearer ${aRandomUniqueString()}")
-
-fun createMockEngine(registry: BehaviorRegistry) = MockEngine { request ->
-    val spec = registry.matchAndPop(request)
-        ?: error("Unexpected request: ${request.method} ${request.url}")
-    respond(
-        content = spec.content,
-        status = spec.status,
-        headers = spec.headers
-    )
-}
